+++ demos/middleware/pages/index.js
@@ -33,6 +33,15 @@ export default function Home() {
             Cookie API
           </Link>
         </p>
+        <p>
+          <Link href="/api/hello">Adds `x-hello` request header</Link>
+        </p>
+        <p>
+          <Link href="/static">Rewrite static page content</Link>
+        </p>
+        <p>
+          <Link href="/headers">Adds `x-hello` request header to a rewrite</Link>
+        </p>
       </main>
     </div>
   )